DEV-1923: Upgrade Cypress to 10
