--- spa/cypress.json
@@ -1,7 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "video": false,
-  "chromeWebSecurity": false,
-  "componentFolder": "src/components",
-  "testFiles": "**/*spec.js"
-}
