+++ spa/cypress/e2e/06 - donation-detail.spec.js
