+++ spa/cypress/e2e/07 - pages.spec.js
