+++ spa/cypress/e2e/08 - contributor.spec.js
