+++ spa/cypress/e2e/09 - styles.spec.js
