+++ spa/cypress/e2e/11 - dashboard.spec.js
