+++ spa/package.json
@@ -26,7 +26,6 @@
     "downshift": "^6.1.3",
     "draft-js": "^0.11.7",
     "draftjs-to-html": "^0.9.1",
-    "find-webpack": "^2.2.1",
     "framer-motion": "^5.3.0",
     "html-to-draftjs": "^1.5.0",
     "html2canvas": "^1.0.0-rc.7",
@@ -70,10 +69,8 @@
     "start": "craco start",
     "start:subdomains": "DANGEROUSLY_DISABLE_HOST_CHECK=true craco start",
     "build": "INLINE_RUNTIME_CHUNK=false craco build",
-    "cypress:open": "cypress open",
-    "cypress:open:component": "cypress open-ct",
-    "cypress:run": "cypress run",
-    "cypress:run:component": "cypress run-ct",
+    "cypress:open": "cypress open --e2e",
+    "cypress:run": "cypress run --e2e",
     "test": "react-scripts test",
     "test:coverage": "CI=true npm test -- --env=jsdom --coverage",
     "test:ci": "CI=true npm test -- --env=jsdom",
@@ -139,9 +136,8 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@cypress/react": "^5.9.4",
-    "@cypress/webpack-dev-server": "1.7.0",
-    "@cypress/webpack-preprocessor": "^5.11.1",
+    "@cypress/react": "^6.2.0",
+    "@cypress/webpack-preprocessor": "^5.12.2",
     "@react-theming/storybook-addon": "^1.1.7",
     "@storybook/addon-actions": "^6.5.10",
     "@storybook/addon-essentials": "^6.5.10",
@@ -159,7 +155,7 @@
     "@testing-library/user-event": "^12.8.3",
     "axios-mock-adapter": "^1.21.2",
     "babel-plugin-named-exports-order": "^0.0.2",
-    "cypress": "^8.5.0",
+    "cypress": "^10.9.0",
     "cypress-localstorage-commands": "^1.7.0",
     "eslint": "^7.25.0",
     "eslint-config-prettier": "^8.3.0",
